requirements.txt
Django==1.11.4
django-rest-framework==0.1.0
djangorestframework==3.6.3
mysql==0.0.1
MySQL-python==1.2.5
pytz==2017.2
wheel==0.24.0
django-datetime-widget==0.9.3